Extract DemoVideoDisplay component to own file
